Need to rework Update Method

+++ src/main/java/io/zipcoder/crudapp/interfaces/PersonRepository.java
@@ -3,7 +3,7 @@ package io.zipcoder.crudapp.interfaces;
 import io.zipcoder.crudapp.Person;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PersonRepository extends CrudRepository<Person,Long> {
+public interface PersonRepository extends CrudRepository<Person,Integer> {
 
 
 }
